my-conversations/requirements.txt
google-cloud-aiplatform==1.66.0
python-dotenv==1.0.1
